@@ -14,47 +14,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/cdk": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/fire": "^5.0.2",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/material": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
+    "@angular/animations": "^7.1.4",
+    "@angular/cdk": "^7.1.4",
+    "@angular/common": "^7.1.4",
+    "@angular/compiler": "^7.1.4",
+    "@angular/core": "^7.1.4",
+    "@angular/fire": "^5.1.1",
+    "@angular/forms": "^7.1.4",
+    "@angular/http": "^7.1.4",
+    "@angular/material": "^7.2.0",
+    "@angular/platform-browser": "^7.1.4",
+    "@angular/platform-browser-dynamic": "^7.1.4",
+    "@angular/router": "^7.1.4",
     "aws-sdk": "^2.377.0",
-    "core-js": "^2.5.7",
-    "firebase": "^5.5.5",
+    "core-js": "^2.6.1",
+    "firebase": "^5.7.2",
     "firebase-admin": "^6.4.0",
     "firebase-functions": "^2.1.0",
     "rxjs": "^6.3.3",
+    "tslib": "^1.9.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.2",
-    "@angular/cli": "~7.0.2",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
-    "@types/jasmine": "~2.8.9",
+    "@angular-devkit/build-angular": "~0.11.4",
+    "@angular/cli": "~7.1.4",
+    "@angular/compiler-cli": "^7.1.4",
+    "@angular/language-service": "^7.1.4",
+    "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.5",
     "@types/node": "~10.12.0",
     "codelyzer": "~4.5.0",
-    "jasmine-core": "~3.2.1",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.4",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.4",
-    "karma-jasmine": "~1.1.2",
+    "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.3.1",
-    "protractor": "~5.4.1",
+    "protractor": "~5.4.2",
     "stylelint": "^9.9.0",
     "stylelint-config-standard": "^18.2.0",
     "ts-node": "~7.0.1",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.3"
+    "tslint": "~5.12.0",
+    "typescript": "~3.1"
   }
 }
